journal - use smaller templates
I should get in the habit of just adding stuff to the journal when I feel like it instead of doing it at the end of the day.
Maybe that's a bad idea though, and doing it at the end of the day means I actually try to do it?
I guess we'll see.
The main reason I want to do this is because I forget what I've eaten in a day really easily.
So I'm going to try to write it down as I do it instead of at the end of the day.
Might change the "dailies" thing back though?